fastapi-blog/requirements.txt
fastapi==0.73.0
uvicorn==0.17.0
pydantic==1.9.0
python-dotenv==0.19.2
sqlalchemy==1.4.31
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
starlette==0.17.1
databases==0.5.5
psycopg2-binary==2.9.3
asyncpg==0.25.0
pydantic[email]
datetime